+++ R/discrete-uniform-distribution.R
@@ -3,7 +3,7 @@
 #' Discrete uniform distribution
 #'
 #' Probability mass function, distribution function, quantile function and random generation
-#' for the Bernoulli distribution.
+#' for the discrete uniform distribution.
 #'
 #' @param x,q	            vector of quantiles.
 #' @param p	              vector of probabilities.
@@ -13,10 +13,14 @@
 #' @param log,log.p	      logical; if TRUE, probabilities p are given as log(p).
 #' @param lower.tail	    logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
 #'                        otherwise, \eqn{P[X > x]}.
-#'                        
+#' 
+#' @details 
+#' 
+#' If \code{min == max}, then discrete uniform distribution is a degenerate distribution.
+#'                                           
 #' @examples 
 #' 
-#' x <- sample.int(10, 1e5, replace = TRUE) 
+#' x <- rdunif(1e5, 1, 10) 
 #' xx <- -1:11
 #' plot(prop.table(table(x)), type = "h")
 #' lines(xx, ddunif(xx, 1, 10), col = "red")