+++ src/const.h
@@ -1,13 +1,18 @@
 
+/*
 #ifndef P_NORM_CONST
 #define P_NORM_CONST 1e+5
 #endif
+*/
 
 #ifndef MIN_DIFF_EPS
 #define MIN_DIFF_EPS 1e-8
 #endif
 
 #ifndef SQRT_2_PI
-#define SQRT_2_PI 2.506628
+#define SQRT_2_PI 2.506628274631000241612
 #endif
 
+#ifndef PHI_0
+#define PHI_0 0.3989422804014327028632
+#endif